Changed code structure to reduce codesmells.
